backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.2
dj-rest-auth==7.0.1
Django==5.2.1
django-allauth==65.8.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
urllib3==2.4.0
